Use long for id sequence and start from 0

+++ app/src/main/java/com/flashsphere/privatednsqs/datastore/IdGenerator.kt
@@ -9,13 +9,13 @@ import kotlinx.coroutines.sync.withLock
 object IdGenerator {
     private val mutex = Mutex()
 
-    suspend fun getNextId(dataStore: DataStore<Preferences>): Int = mutex.withLock {
+    suspend fun getNextId(dataStore: DataStore<Preferences>): Long = mutex.withLock {
         val id = dataStore.get(PreferenceKeys.ID_SEQUENCE) + 1
         dataStore.update(PreferenceKeys.ID_SEQUENCE, id)
         id
     }
 
-    suspend fun getNextId(prefs: MutablePreferences): Int = mutex.withLock {
+    suspend fun getNextId(prefs: MutablePreferences): Long = mutex.withLock {
         val id = (prefs[PreferenceKeys.ID_SEQUENCE.key] ?: PreferenceKeys.ID_SEQUENCE.defaultValue) + 1
         prefs[PreferenceKeys.ID_SEQUENCE.key] = id
         id
